requirements.txt
hues  # Логгирование с указанием времени и цвета

requests  # Синхронные HTTP Запросы

aiohttp # Асинхронные HTTP запросы
peewee # Запросы к БД

aiomysql # Асинхронные запросы к БД
aiopg # Асинхронные запросы к БД
peewee_async # Асинхронные запросы к БД

captcha_solver  # Решение Captcha

langdetect  # Определение языка в плагине озвучивания текста
gTTS  # Озвучивание текста через Google TTS

Pillow  # Работа с изображениями

xmltodict  # Работа с XML
chatterbot # Чат-бот

python-dateutil  # Работа с датами
qrcode  # Работа с QR кодами